Add support for other transports (#6)
Add tls:// support for specifying that an upstream is using
DNS-over-TLS. Create the pieces that allow TLS settings. These can be
reused for grpc (which isn't implemented)

Tested by talking (over TLS) with 9.9.9.9. The following Corefile shows
how this is configured:

~~~
. {
    forward . tls://9.9.9.9 {
        tls_name dns.quad9.net
        health_check 5s
    }
    log
    errors
}
~~~

HC are inbound, so there are using *the same* protocol, in this case
DNS-over-TLS.